Module ic_wasm_utils::instrumentation

This module is responsible for instrumenting wasm binaries on the Internet Computer.

It exports the function instrument which takes a Wasm binary and injects some instrumentation that allows to:

Quantify the amount of execution every function of that module conducts. This quantity is approximated by the sum of cost of instructions executed on the taken execution path.

Verify that no successful memory.grow results in exceeding the available memory allocated to the canister.

Moreover, it exports the function referred to by the start section under the name canister_start and removes the section. (This is needed so that we can run the initialization after we have set the instructions counter to some value).

After the instrumentation, exported functions canister counter_set and canister counter_get can be used to set/get the counter value. Any other function of that module will only be able to execute as long as at every reentrant basic block of its execution path, the counter is verified to be above zero. Otherwise, the function will trap (via calling a special system API call). If the function returns before the counter overflows, the value of the counter is the initial value minus the sum of cost of all executed instructions.

In more details, first, it inserts two System API functions:

(import "__" "out_of_instructions" (func (;0;) (func)))
(import "__" "update_available_memory" (func (;1;) ((param i32 i32) (result i32))))

It then inserts a global mutable counter:

(global (mut i64) (i64.const 0))

and two exported functions setting and reading the instructions value:

(func (;2;) (type 1) (param i64)
  local.get 0
  global.set 1)
(func (;3;) (type 2) (result i64)
  global.get 1)
(export "canister counter_set" (func 2))
(export "canister counter_get" (func 3))

The function canister counter_set should be called before the execution of the instrumented code. After the execution, the counter can be read using the exported function canister counter_get.

Moreover, it injects a decrementation of the instructions counter (by the sum of cost of all instructions inside this block) at the beginning of every non-reentrant block:

global.get 0
i64.const 2
i64.sub
global.set 0

and a decrementation with a counter overflow check at the beginning of every reentrant block (a function or a loop body):

global.get 0
i64.const 8
i64.sub
global.set 0
global.get 0
i64.const 0
i64.lt_s
if  ;; label = @1
  (call x)
end

Note that we omit checking for the counter overflow at the non-reentrant blocks to optimize for performance. The maximal overflow in that case is bound by the length of the longest execution path consisting of non-reentrant basic blocks.

Structs

InstructionCostTable
The metering can be configured by providing a cost-per-instruction table and the default cost for an instruction in case it's not present in the cost table.

InstrumentationOutput

Segments
Vector of heap data chunks with their offsets.

Functions

instrument
Takes a Wasm binary and inserts the instructions metering and memory grow instrumentation.