Function ic_wasm_utils::instrumentation::instrument

pub fn instrument(
    wasm: &BinaryEncodedWasm, 
    instruction_cost_table: &InstructionCostTable
) -> Result<InstrumentationOutput, WasmInstrumentationError>

Takes a Wasm binary and inserts the instructions metering and memory grow instrumentation.

Returns an InstrumentationOutput or an error if the input binary could not be instrumented.